Fixed a check in/out system

+++ ConsidWebExercise/Controllers/CategoryController.cs
@@ -59,18 +59,13 @@ namespace ConsidWebExercise.Controllers
         {
             // Add check if the category is refeered to in any of the library items
             Category category = _db.Categories.Find(id);
-            var libraryItemsWithCategory = _db.LibraryItems.Where(item => item.Category == category).ToList();
-            if(libraryItemsWithCategory.Count() > 0)
-            {
-                return RedirectToAction("Index");
-            }
-            if (category != null)
+            if(_db.LibraryItems.Where(item => item.CategoryId == category.Id).FirstOrDefault() == null && category != null)
             {
                 _db.Categories.Remove(category);
                 _db.SaveChanges();
                 return RedirectToAction("Index");
             }
-            return RedirectToAction("Index");
+            return RedirectToAction();
         }
 
         // POST: Category/Edit/id
